20191112 - Resized
Scaled the size back down to 64 pixels and refactored some of the code
to pull from Creature.DEFAULT_WIDTH so it has less hardcoded values.
Eventually will need to have tile or creature pull from one another, or
have a global variable class like I used to, so that this value only
need to be changed in one place, but will save for after the tutorial.